Добавляет решение задания 'Правда или действие'
