+++ index.html
@@ -128,7 +128,7 @@
                   <input class="text__hashtags" name="hashtags" placeholder="#ХэшТег" maxlength="20" type="text">
                 </div>
                 <div class="img-upload__field-wrapper">
-                  <textarea class="text__description" name="description" placeholder="Ваш комментарий..."></textarea>
+                  <textarea class="text__description" name="description" placeholder="Ваш комментарий..." maxlength="140" type="text"></textarea>
                 </div>
               </fieldset>
